app/requirements-ai.txt
Flask>=2.2,<4
openai>=1.40.0,<2
google-genai>=0.3.0